Adding RootDepth property for moving project root
There is a new section in vs-chromium-project.txt called [Properties]. In there one can add a new property called "RootDepth" which can be set to a positive number. This will move the root of the project forward in the path by the number of directories specified by the number. This can be especially useful for sharing a single vs-chromium-project.txt file across several sub directories.